requirements.txt
aiohttp~=3.8.3
discord.py~=2.1.0